+++ src-js/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+### Added
+- Document-level Form field access and querying via `TextractDocument.form` in addition to `Page.form`.
+- `Page.pageNumber` to find and return 1-based index of the current page in the parent document.
 ### Changed
 - Use new [merged table cells](https://aws.amazon.com/about-aws/whats-new/2022/03/amazon-textract-updates-tables-check-detection/) feature by default, rather than classic split cells.
 - Eliminate trailing whitespace automatically added to Cell.text